cloud_function/functions/requirements.txt
firebase_functions~=0.1.0
firebase-admin>=6.4.0
google-cloud-firestore>=2.14.0
requests>=2.31.0
python-dotenv>=1.0.1